Implemented infinite effects (#6866)
The problem is that we need to ensure an updating ticker for periodic effects to continue
working correctly, but those are typically implemented using the duration ticker.

We don't want to BC break anyone implementing custom effects in PM5, so this is the best
we can do for now.

In this implementation, durations will underflow back to Limits::INT32_MAX when they
would have expired, ensuring a continuous counter. This might cause unexpected behaviour
for consumers of the API, but I don't see another viable solution for PM5.

Co-authored-by: Dylan K. Taylor <dktapps@pmmp.io>