InGamePacketHandler: fixed bridging and block lag
We need to make sure the client's terrain agrees with the server.
However, we also need to make sure we're not sending unnecessary updates,
because the protocol doesn't give us any way to tell the client whether
a particular block update should come before or after its own placement
actions.

When a regular block is placed, syncing the 6 blocks around it causes
air to be sent on the face where the player is likely to place the next
block, which may then arrive on the client after the player has already
placed the next block. This results in flickering as the client's
network catches up, and in newer versions also causes blocks to get
placed on the wrong face somehow.

Since items that place multiple blocks are always given their own special
item IDs (e.g. beds, doors), and such items always have 0 for their
associated blockstate ID on the network, we can safely assume that
any time we see a 0, the client will only have predicted that one
block to be changed, and avoid sending the adjacent update.

fixes #6592
fixes #6810
closes #6803